fix unit test

+++ src/storage/src/storage_failpoints/test_iterator.rs
@@ -416,7 +416,7 @@ async fn test_failpoints_compactor_iterator_recreate() {
         Arc::new(TaskProgress::default()),
         sstable_store,
         100,
-        vec![risingwave_common::catalog::TableId::new(table_id)]
+        vec![risingwave_common::catalog::TableId::new(table_id as u32)]
             .into_iter()
             .collect(),
     );
